Rename panelStack2Tests.tsx
