+++ packages/core/test/index.ts
@@ -58,7 +58,7 @@ import "./multistep-dialog/multistepDialogTests";
 import "./non-ideal-state/nonIdealStateTests";
 import "./overflow-list/overflowListTests";
 import "./overlay/overlayTests";
-import "./panel-stack2/panelStack2Tests";
+import "./panel-stack/panelStackTests";
 import "./popover/popoverTests";
 import "./popover/popperUtilTests";
 import "./portal/portalTests";
